# swagger-spec.json
{"openapi":"3.0.0","paths":{"/token":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/token/mail":{"get":{"operationId":"AppController_sendEmail","parameters":[],"responses":{"200":{"description":""}}}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["User"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}},"tags":["User"]}},"/user/userByCompany/{companyId}":{"get":{"operationId":"UserController_findUserByCompanyId","parameters":[{"name":"companyId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}},"tags":["User"]}},"/user/email/{email}":{"get":{"operationId":"UserController_findUnique","parameters":[{"name":"email","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["User"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["User"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["User"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["User"]}},"/project-material":{"post":{"operationId":"ProjectMaterialController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectMaterialDto"}}}},"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMaterialEntity"}}}}},"tags":["project-material"]},"get":{"operationId":"ProjectMaterialController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["project-material"]}},"/project-material/{id}":{"get":{"operationId":"ProjectMaterialController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMaterialEntity"}}}}},"tags":["project-material"]},"patch":{"operationId":"ProjectMaterialController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectMaterialDto"}}}},"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMaterialEntity"}}}}},"tags":["project-material"]},"delete":{"operationId":"ProjectMaterialController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMaterialEntity"}}}}},"tags":["project-material"]}},"/project-material/projectMaterial/{projectId}":{"get":{"operationId":"ProjectMaterialController_findByProjectId","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectMaterialEntity"}}}}}},"tags":["project-material"]}},"/material":{"post":{"operationId":"MaterialController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMaterialDto"}}}},"responses":{"201":{"description":""}},"tags":["Material"]},"get":{"operationId":"MaterialController_findAll","parameters":[],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MaterialEntity"}}}}}},"tags":["Material"]}},"/material/{id}":{"get":{"operationId":"MaterialController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialEntity"}}}}},"tags":["Material"]},"patch":{"operationId":"MaterialController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMaterialDto"}}}},"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialEntity"}}}}},"tags":["Material"]},"delete":{"operationId":"MaterialController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialEntity"}}}}},"tags":["Material"]}},"/material/materialByCompany/{companyId}":{"get":{"operationId":"MaterialController_findMaterialByCompanyId","parameters":[{"name":"companyId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MaterialEntity"}}}}}},"tags":["Material"]}},"/company":{"post":{"operationId":"CompanyController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyDto"}}}},"responses":{"201":{"description":""}},"tags":["company"]},"get":{"operationId":"CompanyController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["company"]}},"/company/{id}":{"get":{"operationId":"CompanyController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["company"]},"patch":{"operationId":"CompanyController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyDto"}}}},"responses":{"200":{"description":""}},"tags":["company"]},"delete":{"operationId":"CompanyController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["company"]}},"/project":{"post":{"operationId":"ProjectController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":""}},"tags":["project"]},"get":{"operationId":"ProjectController_findAll","parameters":[],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectEntity"}}}}}},"tags":["project"]}},"/project/projectByCompany/{companyId}":{"get":{"operationId":"ProjectController_projectsByCompanyId","parameters":[{"name":"companyId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectEntity"}}}}}},"tags":["project"]}},"/project/{id}":{"get":{"operationId":"ProjectController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectEntity"}}}}},"tags":["project"]},"patch":{"operationId":"ProjectController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectEntity"}}}}},"tags":["project"]},"delete":{"operationId":"ProjectController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectEntity"}}}}},"tags":["project"]}},"/invited-user":{"post":{"operationId":"InvitedUserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInvitedUserDto"}}}},"responses":{"201":{"description":""}},"tags":["invited-user"]},"get":{"operationId":"InvitedUserController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["invited-user"]}},"/invited-user/{id}":{"get":{"operationId":"InvitedUserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["invited-user"]},"patch":{"operationId":"InvitedUserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInvitedUserDto"}}}},"responses":{"200":{"description":""}},"tags":["invited-user"]},"delete":{"operationId":"InvitedUserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["invited-user"]}},"/table-custom-fields":{"post":{"operationId":"TableCustomFieldsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTableCustomFieldDto"}}}},"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableCustomFieldEntities"}}}}},"tags":["table-custom-fields"]},"get":{"operationId":"TableCustomFieldsController_findAll","parameters":[],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TableCustomFieldEntities"}}}}}},"tags":["table-custom-fields"]}},"/table-custom-fields/{id}":{"get":{"operationId":"TableCustomFieldsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableCustomFieldEntities"}}}}},"tags":["table-custom-fields"]},"patch":{"operationId":"TableCustomFieldsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTableCustomFieldDto"}}}},"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableCustomFieldEntities"}}}}},"tags":["table-custom-fields"]},"delete":{"operationId":"TableCustomFieldsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableCustomFieldEntities"}}}}},"tags":["table-custom-fields"]}},"/table-custom-fields/fields/{projectId}":{"get":{"operationId":"TableCustomFieldsController_findCustomFieldsByProjectId","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TableCustomFieldEntities"}}}}}},"tags":["table-custom-fields"]}},"/user-auth":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"isAuthenticate":true,"status":200}}}}}},"tags":["user-auth"]}},"/user-auth/p":{"get":{"operationId":"UserController_gerProducts","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"isAuthenticate":true,"status":200}}}}}},"tags":["user-auth"]}}},"info":{"title":"Easy Cost","description":"Easy Cost API description","version":"1.0","contact":{}},"tags":[{"name":"Easy Cost","description":""}],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"userType":{"type":"string"},"companyId":{"type":"string"},"userId":{"type":"string"},"id":{"type":"string"}},"required":["name","email","userType","companyId","userId","id"]},"UserEntity":{"type":"object","properties":{"email":{"type":"string"},"userType":{"type":"string"},"companyId":{"type":"string"},"id":{"type":"string"},"name":{"type":"string"},"userId":{"type":"string"}},"required":["email","userType","companyId","id","name","userId"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"userType":{"type":"string"},"companyId":{"type":"string"},"userId":{"type":"string"},"id":{"type":"string"}}},"CreateProjectMaterialDto":{"type":"object","properties":{"materialId":{"type":"string"},"projectId":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"quantity":{"type":"number"}},"required":["materialId","projectId","createdAt","quantity"]},"Decimal":{"type":"object","properties":{}},"ProjectMaterialEntity":{"type":"object","properties":{"id":{"type":"string"},"materialId":{"type":"string"},"projectId":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"quantity":{"$ref":"#/components/schemas/Decimal"}},"required":["id","materialId","projectId","createdAt","quantity"]},"UpdateProjectMaterialDto":{"type":"object","properties":{"materialId":{"type":"string"},"projectId":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"quantity":{"type":"number"}}},"CreateMaterialDto":{"type":"object","properties":{"id":{"type":"string"},"materialName":{"type":"string"},"unit":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"userId":{"type":"string"},"co2e":{"type":"number"},"companyId":{"type":"string"},"hourPerQuantity":{"type":"number"},"price":{"type":"number"}},"required":["id","materialName","unit","createdAt","userId","co2e","companyId","hourPerQuantity","price"]},"MaterialEntity":{"type":"object","properties":{"price":{"type":"number"},"hourPerQuantity":{"type":"number"},"id":{"type":"string"},"materialName":{"type":"string"},"unit":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"pricePerHour":{"$ref":"#/components/schemas/Decimal"},"userId":{"type":"string"},"co2e":{"type":"number"},"companyId":{"type":"string"}},"required":["price","hourPerQuantity","id","materialName","unit","createdAt","pricePerHour","userId","co2e","companyId"]},"UpdateMaterialDto":{"type":"object","properties":{"id":{"type":"string"},"materialName":{"type":"string"},"unit":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"userId":{"type":"string"},"co2e":{"type":"number"},"companyId":{"type":"string"},"hourPerQuantity":{"type":"number"},"price":{"type":"number"}}},"CreateCompanyDto":{"type":"object","properties":{"name":{"type":"string"},"country":{"type":"string"},"logo":{"type":"string"}},"required":["name","country","logo"]},"UpdateCompanyDto":{"type":"object","properties":{"name":{"type":"string"},"country":{"type":"string"},"logo":{"type":"string"}}},"CreateProjectDto":{"type":"object","properties":{"id":{"type":"string"},"projectName":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"userId":{"type":"string"},"location":{"type":"string"},"companyId":{"type":"string"},"isActive":{"type":"boolean"}},"required":["id","projectName","createdAt","userId","location","companyId","isActive"]},"ProjectEntity":{"type":"object","properties":{"id":{"type":"string"},"projectName":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"userId":{"type":"string"},"location":{"type":"string"},"companyId":{"type":"string"},"isActive":{"type":"boolean"}},"required":["id","projectName","createdAt","userId","location","companyId","isActive"]},"UpdateProjectDto":{"type":"object","properties":{"id":{"type":"string"},"projectName":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"userId":{"type":"string"},"location":{"type":"string"},"companyId":{"type":"string"},"isActive":{"type":"boolean"}}},"CreateInvitedUserDto":{"type":"object","properties":{"companyId":{"type":"string"},"userId":{"type":"string"},"projectId":{"type":"string"},"token":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}},"required":["companyId","userId","projectId","token","email","role"]},"UpdateInvitedUserDto":{"type":"object","properties":{"companyId":{"type":"string"},"userId":{"type":"string"},"projectId":{"type":"string"},"token":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}}},"CreateTableCustomFieldDto":{"type":"object","properties":{"id":{"type":"string"},"projectId":{"type":"string"},"columnName":{"type":"string"},"columnType":{"type":"string"},"customFieldValue":{"type":"object"}},"required":["id","projectId","columnName","columnType","customFieldValue"]},"TableCustomFieldEntities":{"type":"object","properties":{"id":{"type":"string"},"projectId":{"type":"string"},"columnName":{"type":"string"},"columnType":{"type":"string"},"customFieldValue":{"type":"object"}},"required":["id","projectId","columnName","columnType","customFieldValue"]},"UpdateTableCustomFieldDto":{"type":"object","properties":{"id":{"type":"string"},"projectId":{"type":"string"},"columnName":{"type":"string"},"columnType":{"type":"string"},"customFieldValue":{"type":"object"}}},"UserDTO":{"type":"object","properties":{"displayName":{"type":"string","description":"displayName"},"email":{"type":"string","description":"email"},"password":{"type":"string","description":"password"},"role":{"type":"string","description":"role"}},"required":["displayName","email","password","role"]}}}}